Started the homework assignment.
In this assignment, I am creating a daily planner
for normal business hours.

At this moment, I have added the first timeblock
for 9am. Saving now as this particular timeblock
is working the way I want it to as of now
and I want to clean up the javascript and then
get ready to duplicate the code for the other hours